+++ src/test/java/com/maxmind/minfraud/response/DeviceTest.java
@@ -19,7 +19,7 @@ public class DeviceTest extends AbstractOutputTest {
                         .put("confidence", 99)
                         .put("id", "C8D3BE1A-BE26-11E5-8C50-1B575C37265F")
                         .put("last_seen", "2016-06-08T14:16:38Z")
-                        .put("local_time", "2018-04-05T15:21:00-07:00")
+                        .put("local_time", "2018-04-05T15:21:01-07:00")
                         .end()
                         .finish()
         );
@@ -28,8 +28,8 @@ public class DeviceTest extends AbstractOutputTest {
         assertEquals(UUID.fromString("C8D3BE1A-BE26-11E5-8C50-1B575C37265F"), device.getId());
         assertEquals("2016-06-08T14:16:38Z", device.getLastSeen());
         assertEquals("2016-06-08T14:16:38Z", device.getLastSeenDateTime().toString());
-        assertEquals("2018-04-05T15:21:00-07:00", device.getLocalTime());
-        assertEquals("2018-04-05T15:21:00-07:00", device.getLocalDateTime().toString());
+        assertEquals("2018-04-05T15:21:01-07:00", device.getLocalTime());
+        assertEquals("2018-04-05T15:21:01-07:00", device.getLocalDateTime().toString());
 
     }
 }